abstracted to avoid future merge conflicts + cleaner
