+++ crab/tree.go
@@ -884,7 +884,9 @@ func (t *CrabTree) deleteEntry(n *Node, key int, pointer interface{}) {
 
 }
 
-func (t *CrabTree) Stage1(Q []tree_api.Query, i int, num_threads int) {}
-func (t *CrabTree) Stage2(Q []tree_api.Query, i int, num_threads int) {}
-func (t *CrabTree) Stage3(Q []tree_api.Query, i int, num_threads int) {}
-func (t *CrabTree) Stage4(Q []tree_api.Query, i int, num_threads int) {}
+func (t *CrabTree) Palm(key_count int, num_threads int) {}
+
+// func (t *CrabTree) Stage1(Q []tree_api.Query, i int, num_threads int) {}
+// func (t *CrabTree) Stage2(Q []tree_api.Query, i int, num_threads int) {}
+// func (t *CrabTree) Stage3(Q []tree_api.Query, i int, num_threads int) {}
+// func (t *CrabTree) Stage4(Q []tree_api.Query, i int, num_threads int) {}
