+++ lock_free/stage2.go
@@ -1,0 +1,13 @@
+package lock_free
+
+import (
+	// "errors"
+	// "fmt"
+	"main/tree_api"
+	// "reflect"
+	// "sync"
+)
+
+func (t *LockFreeTree) Stage2(sharedLeafData [][]*Node) ([](map[*Node]([]Modification)), []*tree_api.Record) {
+	return nil, nil
+}
