+++ lock_free/stage3.go
@@ -1,0 +1,13 @@
+package lock_free
+
+// import (
+// "errors"
+// "fmt"
+// "main/tree_api"
+// "reflect"
+// "sync"
+// )
+
+func (t *LockFreeTree) Stage3(sharedModLists [](map[*Node]([]Modification))) [](map[*Node]([]Modification)) {
+	return nil
+}
