+++ lock_free/stage4.go
@@ -1,0 +1,11 @@
+package lock_free
+
+// import (
+// 	"errors"
+// 	"fmt"
+// 	"main/tree_api"
+// 	"reflect"
+// 	"sync"
+// )
+
+func (t *LockFreeTree) Stage4(finalModList [](map[*Node]([]Modification))) {}
