+++ seq_tree/tree.go
@@ -801,7 +801,9 @@ func (t *Tree) deleteEntry(n *Node, key int, pointer interface{}) {
 
 }
 
-func (t *Tree) Stage1(Q []tree_api.Query, i int, num_threads int) {}
-func (t *Tree) Stage2(Q []tree_api.Query, i int, num_threads int) {}
-func (t *Tree) Stage3(Q []tree_api.Query, i int, num_threads int) {}
-func (t *Tree) Stage4(Q []tree_api.Query, i int, num_threads int) {}
+func (t *Tree) Palm(key_count int, num_threads int) {}
+
+// func (t *Tree) Stage1(Q []tree_api.Query, i int, num_threads int) {}
+// func (t *Tree) Stage2(Q []tree_api.Query, i int, num_threads int) {}
+// func (t *Tree) Stage3(Q []tree_api.Query, i int, num_threads int) {}
+// func (t *Tree) Stage4(Q []tree_api.Query, i int, num_threads int) {}
