+++ tree_api/tree_api.go
@@ -50,8 +50,9 @@ type BPTree interface {
 	// FindAndPrint(key int, verbose bool)
 	// FindAndPrintRange(key_start, key_end int, verbose bool)
 	// Stage1(Q []Query, i int, num_threads int)
-	Stage1(Q []Query, i int, num_threads int) interface{}
-	Stage2(Q []Query, i int, num_threads int)
-	Stage3(Q []Query, i int, num_threads int)
-	Stage4(Q []Query, i int, num_threads int)
+	Palm(key_count int, num_threads int)
+	// Stage1(Q []Query, i int, num_threads int)
+	// Stage2(Q []Query, i int, num_threads int)
+	// Stage3(Q []Query, i int, num_threads int)
+	// Stage4(Q []Query, i int, num_threads int)
 }
